IChartDataAdapter

+++ AndroidChart/app/src/main/java/z/cn/chart/adapter/IMapChartAdapter.java
@@ -3,13 +3,17 @@ package z.cn.chart.adapter;
 import java.util.Map;
 
 /**
- * Created by rui on 2017/1/12.
+ * Created by isee15 on 2017/1/12.
  */
 
 public interface IMapChartAdapter {
     int getResourceId();
+
     int[] getColorRange();
+
     int getValueMin();
+
     int getValueMax();
-    Map<String,Integer> getValues();
+
+    Map<String, Double> getValues();
 }
